creación de Policy isAdmin, método delete y tests

+++ tienda-app/app/Http/Controllers/UserController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Redirect;
 use Inertia\Response;
 use Inertia\Inertia;
 
@@ -78,8 +80,11 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(User $user)
+    public function destroy(User $user): RedirectResponse
     {
+        Gate::authorize('deleteUsuario', $user);
         $user->delete();
+
+        return redirect(route('usuarios.index'))->with('success', 'El usuario ha sido correctamente eliminado.');
     }
 }
